Fixed decoder initialization. fixes #30
This change ensures that decoders and their internals are properly
initialized when being constructed. This also removes some undefined
behavior where the inner field of an uninitialized decoder was being
accessed during construction. Unfortunately it requires a small API
change but now the Box/Arc of Decoders and SyncDecoders is internal and
no longer has to appear in the public API.